requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --cert=None --client-cert=None --index-url=None --pip-args=None requirements.in
#
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.13.1
    # via edge-tts
aiosignal==1.4.0
    # via aiohttp
alembic==1.16.5
    # via -r requirements.in
amqp==5.3.1
    # via kombu
annotated-types==0.7.0
    # via pydantic
anyio==4.11.0
    # via
    #   groq
    #   httpx
    #   starlette
    #   watchfiles
asyncpg==0.30.0
    # via -r requirements.in
attrs==25.4.0
    # via aiohttp
audioread==3.0.1
    # via librosa
bcrypt==4.0.1
    # via -r requirements.in
billiard==4.2.2
    # via celery
celery==5.4.0
    # via -r requirements.in
certifi==2025.10.5
    # via
    #   edge-tts
    #   httpcore
    #   httpx
    #   requests
cffi==2.0.0
    # via
    #   cryptography
    #   gevent
    #   soundfile
charset-normalizer==3.4.3
    # via requests
click==8.3.0
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   uvicorn
click-didyoumean==0.3.1
    # via celery
click-plugins==1.1.1.2
    # via celery
click-repl==0.3.0
    # via celery
colorama==0.4.6
    # via
    #   click
    #   uvicorn
coloredlogs==15.0.1
    # via onnxruntime
cryptography==46.0.2
    # via -r requirements.in
decorator==5.2.1
    # via librosa
distro==1.9.0
    # via groq
dnspython==2.8.0
    # via email-validator
edge-tts==7.2.3
    # via -r requirements.in
email-validator==2.3.0
    # via pydantic
fastapi==0.118.0
    # via -r requirements.in
ffmpeg-python==0.2.0
    # via -r requirements.in
flatbuffers==25.9.23
    # via onnxruntime
frozenlist==1.8.0
    # via
    #   aiohttp
    #   aiosignal
future==1.0.0
    # via ffmpeg-python
gevent==25.9.1
    # via -r requirements.in
greenlet==3.2.4
    # via
    #   gevent
    #   sqlalchemy
groq==0.32.0
    # via -r requirements.in
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httptools==0.6.4
    # via uvicorn
httpx==0.28.1
    # via groq
humanfriendly==10.0
    # via coloredlogs
idna==3.10
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   requests
    #   yarl
joblib==1.5.2
    # via
    #   librosa
    #   scikit-learn
kombu==5.5.4
    # via celery
lazy-loader==0.4
    # via librosa
librosa==0.11.0
    # via -r requirements.in
llvmlite==0.45.1
    # via numba
mako==1.3.10
    # via alembic
markupsafe==3.0.3
    # via mako
mpmath==1.3.0
    # via sympy
msgpack==1.1.1
    # via librosa
multidict==6.7.0
    # via
    #   aiohttp
    #   yarl
numba==0.62.1
    # via librosa
numpy==2.3.3
    # via
    #   librosa
    #   numba
    #   onnxruntime
    #   scikit-learn
    #   scipy
    #   soundfile
    #   soxr
onnxruntime==1.23.0
    # via piper-tts
packaging==25.0
    # via
    #   kombu
    #   lazy-loader
    #   onnxruntime
    #   pooch
passlib==1.7.4
    # via -r requirements.in
piper-tts==1.3.0
    # via -r requirements.in
platformdirs==4.4.0
    # via pooch
pooch==1.8.2
    # via librosa
prompt-toolkit==3.0.52
    # via click-repl
propcache==0.4.1
    # via
    #   aiohttp
    #   yarl
protobuf==6.32.1
    # via onnxruntime
psycopg2-binary==2.9.10
    # via -r requirements.in
pycparser==2.23
    # via cffi
pydantic[email]==2.11.10
    # via
    #   -r requirements.in
    #   fastapi
    #   groq
    #   pydantic-settings
pydantic-core==2.33.2
    # via pydantic
pydantic-settings==2.11.0
    # via -r requirements.in
pydub==0.25.1
    # via -r requirements.in
pyreadline3==3.5.4
    # via humanfriendly
python-dateutil==2.9.0.post0
    # via celery
python-dotenv==1.1.1
    # via
    #   -r requirements.in
    #   pydantic-settings
    #   uvicorn
python-multipart==0.0.20
    # via -r requirements.in
pyyaml==6.0.3
    # via uvicorn
redis==5.0.4
    # via -r requirements.in
requests==2.32.5
    # via pooch
scikit-learn==1.7.2
    # via librosa
scipy==1.16.2
    # via
    #   librosa
    #   scikit-learn
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   groq
soundfile==0.13.1
    # via
    #   -r requirements.in
    #   librosa
soxr==1.0.0
    # via librosa
sqlalchemy[asyncio]==2.0.43
    # via
    #   -r requirements.in
    #   alembic
starlette==0.48.0
    # via fastapi
sympy==1.14.0
    # via onnxruntime
tabulate==0.9.0
    # via edge-tts
threadpoolctl==3.6.0
    # via scikit-learn
typing-extensions==4.15.0
    # via
    #   aiosignal
    #   alembic
    #   anyio
    #   edge-tts
    #   fastapi
    #   groq
    #   librosa
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   starlette
    #   typing-inspection
typing-inspection==0.4.2
    # via
    #   pydantic
    #   pydantic-settings
tzdata==2025.2
    # via
    #   celery
    #   kombu
urllib3==2.5.0
    # via requests
uvicorn[standard]==0.37.0
    # via -r requirements.in
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
watchfiles==1.1.0
    # via uvicorn
wcwidth==0.2.14
    # via prompt-toolkit
websockets==15.0.1
    # via uvicorn
yarl==1.22.0
    # via aiohttp
zope-event==6.0
    # via gevent
zope-interface==8.0.1
    # via gevent

# The following packages are considered to be unsafe in a requirements file:
# setuptools
